+++ packages/runtime/src/helpers/config.ts
@@ -6,7 +6,7 @@ import type { NextConfigComplete } from 'next/dist/server/config-shared'
 import { join, dirname, relative } from 'pathe'
 import slash from 'slash'
 
-import { HANDLER_FUNCTION_NAME, ODB_FUNCTION_NAME } from '../constants'
+import { HANDLER_FUNCTION_NAME, IMAGE_FUNCTION_NAME, ODB_FUNCTION_NAME } from '../constants'
 
 import type { RoutesManifest } from './types'
 import { escapeStringRegexp } from './utils'
@@ -71,7 +71,7 @@ export const updateRequiredServerFiles = async (publish: string, modifiedConfig:
   await writeJSON(configFile, modifiedConfig)
 }
 
-const resolveModuleRoot = (moduleName) => {
+export const resolveModuleRoot = (moduleName) => {
   try {
     return dirname(relative(process.cwd(), require.resolve(`${moduleName}/package.json`, { paths: [process.cwd()] })))
   } catch {
@@ -107,10 +107,9 @@ export const configureHandlerFunctions = async ({
   const files = config.files || []
   const cssFilesToInclude = files.filter((f) => f.startsWith(`${publish}/static/css/`))
 
-  /* eslint-disable no-underscore-dangle */
   if (!destr(process.env.DISABLE_IPX)) {
-    netlifyConfig.functions._ipx ||= {}
-    netlifyConfig.functions._ipx.node_bundler = 'nft'
+    netlifyConfig.functions[IMAGE_FUNCTION_NAME] ||= {}
+    netlifyConfig.functions[IMAGE_FUNCTION_NAME].node_bundler = 'nft'
   }
 
   // If the user has manually added the module to included_files, then don't exclude it
@@ -118,7 +117,6 @@ export const configureHandlerFunctions = async ({
     (moduleName) => !hasManuallyAddedModule({ netlifyConfig, moduleName }),
   )
 
-  /* eslint-enable no-underscore-dangle */
   ;[HANDLER_FUNCTION_NAME, ODB_FUNCTION_NAME, '_api_*'].forEach((functionName) => {
     netlifyConfig.functions[functionName] ||= { included_files: [], external_node_modules: [] }
     netlifyConfig.functions[functionName].node_bundler = 'nft'